pyproject.toml
[tool.poetry]
name = "strawberry-tutorial"
version = "0.1.0"
description = ""
authors = ["jasonxie <x5758x@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = {extras = ["debug-server"], version = "^0.205.0"}
faker = "19.6.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.15.0"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.287"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
